Rename 'raw' to 'snapshot'
