+++ backend/src/domain/inventory/models/index.ts
@@ -1,1 +1,1 @@
-export * from "./raw";
+export * from "./snapshots";
