+++ backend/src/domain/inventory/models/snapshots/cabinetItem.ts
