+++ backend/src/domain/inventory/models/snapshots/index.ts
