+++ backend/src/domain/inventory/models/snapshots/product.ts
